Deploy current production Wordpress locally
This set of changes gets an vagrant/ansible local instance
of our MITH website running. The database dump needed to be
rewritten to change mith.umd.edu to localhost:8080 since multisite
Wordpress installations are a bit trickier to move around. In
addition I needed to disable the block-bad-queries plugin since
it interfered with some URLs having localhost in them. Lastly
I needed to modify wordpress.conf to allow php-fpm to execute
source code in /vagrant, which is where wp-content is made available
to easy editing by the Vagrant host.