simplified instructions
This directory contains a generic wordpress playbook as well as
playbooks to fetch MITH Wordpress data from a remote machine and
to setup a dev instance of MITH Wordpress locally using Vagrant.

At some point it might make sense to disentangle these, but for now
they live together. I simplified the README.md for now to make it
easier for someone to get up and running with the dev instance of
MITH Wordpress without needing to understand all these details.